dumbdec, baseclass: initial subsong support working
dumbdec: moved handle_loop() call back to where it was (supposed
improvement was a bug in disguise)

TODO: optional loop and subsong properties aren't working out. Always
set them in the base class, even if subclasses do not support subsongs
and/or looping. An unfortunate limitation.. (once this is done, reset
PROP_0 in dumb_dec to 0)

also, the extra code for getting the subsongs probably should be part of
DUMB instead.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>