changed subsong handling in order to be able to create a TOC
Before, it was up to the subclasses to update the duration when a
new subsong was to be played. This is done by the base class now.
The base class retrieves information (= duration and tags) for all
subsongs right after load_from_buffer() was called.

Two new class functions were introduced for subsong duration&tags,
set_duration() is gone, the num-subsongs property is also gone.

Base class also sends tag information for the initial subsong
automatically, as well as after switching to a new subsong. Subclasses
are still free to send additional tags.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>