Avoid overwriting tag lists which haven't been pushed yet
Before, it was possible that some tags never made it downstream. For
example, if a subclass created tags & pushed a tag event downstream as
part of its load_from_buffer() function, and then later the tags related
to the subsong duration are pushed downstream, then the former tags will
be overwritten by the latter if the former haven't been pushed further
yet (for example, in decodebin, because the pads aren't linked yet).
Avoid this by always merging subsong tags with main tags.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>