baseclass: Improve gst_nonstream_audio_decoder_get_downstream_info()
Instead of normalizing caps and picking the first match, look at the
format, rate, samples, and try to fixate them. This also makes sure that
even if for example the channels caps is set to {1,2,4}, and the default
channel count is 2, it is fixated to 2, making the nonstandard order of
format and channel count in the openmptdec and dumbdec template caps
unnecessary.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>